Test support for long records.

+++ lib/textfile.rb
@@ -5,20 +5,25 @@ class Textfile < Pathname
   # See http://apple.stackexchange.com/questions/69223/how-to-replace-mac-os-x-utilities-with-gnu-core-utilities
   COMM_CMD = (RUBY_PLATFORM =~ /darwin/ ? 'gcomm' : 'comm')
   SORT_CMD = (RUBY_PLATFORM =~ /darwin/ ? 'gsort' : 'sort')
+  UNIQ_CMD = (RUBY_PLATFORM =~ /darwin/ ? 'guniq' : 'uniq')
 
-  def initialize(p1)#, options = {})
+  def initialize(p1, options = {})
+    @bufsiz = options[:bufsiz]
+    @debug = options[:debug]
     super(p1)
   end
 
   # Sorts file and removes any duplicate records.
   def sort(options='')
+    options.concat(" --buffer-size=#{@bufsiz}") if @bufsiz
     with_tempcopy do |tempcopy|
-      sh "#{SORT_CMD} #{tempcopy} | uniq > #{self}"
+      sh "#{SORT_CMD} #{options} #{tempcopy} | #{UNIQ_CMD} > #{self}"
     end
   end
 
   protected
   def sh(cmd)
+    puts cmd if @debug
     %x[ #{cmd} ]
     self
   end
@@ -28,7 +33,7 @@ class Textfile < Pathname
     tempcopy.write(self.read)
     tempcopy.close
     yield tempcopy.path
-    tempcopy.unlink
+    tempcopy.unlink unless @debug
     self
   end
 end
